Backend/requirements.txt
fastapi
uvicorn
python-dotenv
requests
groq
yfinance
sqlalchemy
pydantic
websockets